+++ packages/core/__tests__/helpers/app.tsx
@@ -6,7 +6,7 @@ import { Container, ContainerProvider, TOKENS } from '../../src/container-api'
 import { MainContainer } from '../../src/container-impl'
 import { container } from 'tsyringe'
 import { OpenIDCredentialRecordProvider } from '../../src/modules/openid/context/OpenIDCredentialRecordProvider'
-import { MockLogger } from './logger'
+import { MockLogger } from '../../src/testing/MockLogger'
 
 export const BasicAppContext: React.FC<PropsWithChildren> = ({ children }) => {
   const context = useMemo(() => {