+++ packages/core/src/components/misc/CredentialCard.tsx
@@ -8,12 +8,13 @@ import { useTheme } from '../../contexts/theme'
 import { GenericFn } from '../../types/fn'
 
 import CredentialCard10 from './CredentialCard10'
-import CredentialCard11, { CredentialErrors } from './CredentialCard11'
-import { GenericCredentialExchangeRecord } from '../../types/credentials'
+import CredentialCard11 from './CredentialCard11'
+import { CredentialErrors, GenericCredentialExchangeRecord } from '../../types/credentials'
 import { BrandingOverlay } from '@bifold/oca'
 import { useOpenIDCredentials } from '../../modules/openid/context/OpenIDCredentialRecordProvider'
 import { getCredentialForDisplay } from '../../modules/openid/display'
 import { getAttributeField } from '../../utils/oca'
+import { useCredentialErrorsFromRegistry } from '../../modules/openid/hooks/useCredentialErrorsFromRegistry'
 
 interface CredentialCardProps {
   credential?: GenericCredentialExchangeRecord
@@ -50,6 +51,7 @@ const CredentialCard: React.FC<CredentialCardProps> = ({
   const [overlay, setOverlay] = useState<CredentialOverlay<BrandingOverlay>>({})
   const { resolveBundleForCredential } = useOpenIDCredentials()
   const [extraOverlayAttribute, setExtraOverlayAttribute] = useState<Attribute | undefined>()
+  const computedErrors = useCredentialErrorsFromRegistry(credential, credentialErrors)
 
   useEffect(() => {
     if (brandingOverlay) {
@@ -138,7 +140,7 @@ const CredentialCard: React.FC<CredentialCardProps> = ({
         style={style}
         onPress={onPress}
         brandingOverlay={overlay}
-        credentialErrors={credentialErrors ?? []}
+        credentialErrors={computedErrors}
         proof={proof}
         elevated={proof}
         displayItems={displayItems}