+++ packages/core/src/modules/openid/offerResolve.tsx
@@ -23,7 +23,11 @@ import {
   W3cJwtVerifiableCredential,
   Buffer as CredoBuffer,
 } from '@credo-ts/core'
-import { extractOpenId4VcCredentialMetadata, setOpenId4VcCredentialMetadata } from './metadata'
+import {
+  extractOpenId4VcCredentialMetadata,
+  setOpenId4VcCredentialMetadata,
+  temporaryMetaVanillaObject,
+} from './metadata'
 import Config from 'react-native-config'
 
 export const resolveOpenId4VciOffer = async ({
@@ -77,7 +81,7 @@ export async function acquirePreAuthorizedAccessToken({
   agent: Agent
   resolvedCredentialOffer: OpenId4VciResolvedCredentialOffer
   txCode?: string
-}) {
+}): Promise<OpenId4VciRequestTokenResponse> {
   return await agent.modules.openId4VcHolder.requestToken({
     resolvedCredentialOffer,
     txCode,
@@ -187,14 +191,14 @@ export const customCredentialBindingResolver = async ({
 export const receiveCredentialFromOpenId4VciOffer = async ({
   agent,
   resolvedCredentialOffer,
-  accessToken,
+  tokenResponse,
   credentialConfigurationIdsToRequest,
   clientId,
   pidSchemes,
 }: {
   agent: Agent
   resolvedCredentialOffer: OpenId4VciResolvedCredentialOffer
-  accessToken: OpenId4VciRequestTokenResponse
+  tokenResponse: OpenId4VciRequestTokenResponse
   credentialConfigurationIdsToRequest?: string[]
   clientId?: string
   pidSchemes?: { sdJwtVcVcts: Array<string>; msoMdocDoctypes: Array<string> }
@@ -213,7 +217,7 @@ export const receiveCredentialFromOpenId4VciOffer = async ({
 
   const credentials = await agent.modules.openId4VcHolder.requestCredentials({
     resolvedCredentialOffer,
-    ...accessToken,
+    ...tokenResponse,
     clientId,
     credentialsToRequest: credentialConfigurationIdsToRequest,
     verifyCredentialStatus: false,
@@ -248,6 +252,7 @@ export const receiveCredentialFromOpenId4VciOffer = async ({
 
   // We only support one credential for now
   const [firstCredential] = credentials
+
   if (!firstCredential)
     throw new Error('Error retrieving credential using pre authorized flow: firstCredential undefined!.')
 
@@ -272,6 +277,11 @@ export const receiveCredentialFromOpenId4VciOffer = async ({
       // We don't support expanded types right now, but would become problem when we support JSON-LD
       tags: {},
     })
+  }
+
+  const notificationMetadata = { ...firstCredential.notificationMetadata }
+  if (notificationMetadata) {
+    temporaryMetaVanillaObject.notificationMetadata = notificationMetadata
   }
 
   const openId4VcMetadata = extractOpenId4VcCredentialMetadata(